historical/requirements.txt
python-binance
requests
python-dotenv
pandas
elasticsearch
